pyproject.toml
[project]
name = "trading-v3"
version = "0.1.0"
description = "DRL-based crypto trading bot using TD3"
requires-python = ">=3.12"
dependencies = [
    "ccxt>=4.0",
    "gymnasium>=0.29",
    "stable-baselines3>=2.3",
    "pandas>=2.1",
    "numpy>=1.26",
    "pandas-ta>=0.3.14",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "mlflow>=2.12",
    "torch>=2.1",
    "matplotlib>=3.8",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4",
    "pytest>=8.0",
    "pytest-mock>=3.14",
    "ipykernel>=6.29",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
